Description
As a data analyst for a European sports clothing company, I was tasked with determining which of the Gulf Coastal Countries would be the best to expand into. To do this, I analyzed several datasets containing relevant information and provided a structured report for senior leadership.

My report was divided into several sections, starting with a breakdown of the business problem. I then presented my findings and recommendations based on the data analysis, along with any assumptions made and data quality and availability constraints.

To arrive at my recommendations, I conducted a detailed analysis of the data provided and explained how I derived the answers. All code used in the analysis was written in Python, and the report was presented in a Jupyter Notebook format.

Ultimately, my goal was to provide senior leadership with a well-informed recommendation on which Gulf Coastal Country would be the most promising for expansion. Through my analysis and recommendations, I was able to help guide the company towards making a strategic decision on its international expansion plans.